refactor(restify): split EndpointParser into OSGi parser services
